+++ frontend/src/components/PageHeader/types.d.ts
@@ -1,0 +1,12 @@
+interface IPageHeader {
+    pageTitle: string;
+}
+
+interface IBreadcrumbPageHeader extends IPageHeader {
+    breadcrumbItems: string[];
+}
+
+interface ISubPageHeader extends IPageHeader {
+    chevronLink: string;
+    children: React.ReactNode;
+}
