--- frontend/src/components/PageHeader/types.ts
@@ -1,3 +1,0 @@
-export interface IPageHeader {
-    pageTitle: string;
-}
